ChatBridge/requirements.txt
accelerate==0.16.0
aiofiles==24.1.0
aiohttp==3.8.4
aiosignal==1.3.1
altair==5.5.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.6.2.post1
asttokens==3.0.0
async-lru==2.0.4
async-timeout==4.0.2
attrs==22.2.0
babel==2.16.0
beautifulsoup4==4.12.3
bitsandbytes==0.37.0
bleach==6.2.0
blis==0.7.11
braceexpand==0.1.7
Brotli==1.0.9
catalogue==2.0.10
cchardet==2.1.7
certifi==2024.8.30
cffi==1.17.1
chardet==5.1.0
charset-normalizer==3.3.2
click==8.1.7
comm==0.2.2
confection==0.1.5
contourpy==1.0.7
cycler==0.11.0
cymem==2.0.10
debugpy==1.8.9
decorator==5.2.1
decord==0.6.0
defusedxml==0.7.1
distlib==0.3.9
docker-pycreds==0.4.0
easydict==1.13
exceptiongroup==1.2.2
executing==2.2.0
faiss-gpu==1.7.2
fastapi==0.115.5
fastjsonschema==2.21.1
ffmpy==0.4.0
filelock==3.16.1
fonttools==4.38.0
fqdn==1.5.1
frozenlist==1.3.3
fsspec==2024.10.0
gitdb==4.0.11
GitPython==3.1.43
gradio==3.24.1
gradio_client==0.0.8
h11==0.14.0
h5py==3.12.1
httpcore==1.0.7
httpx==0.28.0
huggingface-hub==0.13.4
idna==3.7
importlib-resources==5.12.0
importlib_metadata==8.4.0
intel-cmplr-lib-ur==2025.0.4
intel-openmp==2025.0.4
iopath==0.1.10
ipdb==0.13.13
ipython==8.18.1
jedi==0.19.2
Jinja2==3.1.4
joblib==1.4.2
json5==0.10.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
kiwisolver==1.4.4
langcodes==3.5.0
language_data==1.3.0
linkify-it-py==2.0.3
llvmlite==0.43.0
marisa-trie==1.2.1
markdown-it-py==2.2.0
MarkupSafe==3.0.2
matplotlib==3.7.0
matplotlib-inline==0.1.7
mdit-py-plugins==0.3.3
mdurl==0.1.2
mistune==3.0.2
mkl==2025.0.1
mkl-service==2.4.0
mkl_fft==1.3.11
mkl_random==1.2.8
multidict==6.0.4
murmurhash==1.0.11
narwhals==1.15.1
nltk==3.9.1
numba==0.60.0
numpy==1.26.4
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==9.7.1.26
nvidia-nccl-cu11==2.21.5
omegaconf==2.3.0
openai==0.27.0
opencv-python==4.7.0.72
orjson==3.10.12
overrides==7.7.0
packaging==23.0
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pathlib_abc==0.1.1
pathy==0.11.0
pexpect==4.9.0
pillow==11.0.0
platformdirs==4.3.6
portalocker==3.0.0
preshed==3.0.9
prometheus_client==0.21.0
prompt_toolkit==3.0.48
propcache==0.3.0
protobuf==5.29.0
psutil==5.9.4
ptyprocess==0.7.0
pure_eval==0.2.3
pycocoevalcap==1.2
pycocotools==2.0.6
pycparser==2.22
pydantic==1.10.19
pydantic_core==2.27.2
pydub==0.25.1
Pygments==2.18.0
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
python-json-logger==2.0.7
python-multipart==0.0.19
pytz==2024.2
PyYAML==6.0
pyzmq==26.2.0
referencing==0.35.1
regex==2022.10.31
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.21.0
safetensors==0.4.5
scikit-learn==1.2.2
scipy==1.10.1
semantic-version==2.10.0
Send2Trash==1.8.3
sentencepiece==0.2.0
sentry-sdk==2.19.0
setproctitle==1.3.4
six==1.16.0
smart-open==6.4.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
spacy==3.5.1
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
stack-data==0.6.3
starlette==0.41.3
tbb==2022.0.0
tcmlib==1.2.0
tenacity==8.2.2
terminado==0.18.1
thinc==8.1.12
threadpoolctl==3.5.0
timm==0.6.13
tinycss2==1.4.0
tokenizers==0.13.2
tomli==2.2.1
toolz==1.0.0
torch==1.12.0+cu116
torchaudio==0.12.0+cu116
torchvision==0.13.0+cu116
tornado==6.4.2
tqdm==4.64.1
traitlets==5.14.3
transformers==4.28.0
typer==0.7.0
types-python-dateutil==2.9.0.20241003
typing_extensions==4.11.0
tzdata==2024.2
uc-micro-py==1.0.3
umf==0.9.1
uri-template==1.3.0
urllib3==2.2.3
uvicorn==0.32.1
virtualenv==20.28.1
wandb==0.18.7
wasabi==1.1.3
wcwidth==0.2.13
webcolors==24.11.1
webdataset==0.2.48
webencodings==0.5.1
websocket-client==1.8.0
websockets==12.0
yarl==1.8.2
zipp==3.14.0
